Servo.py Updated initalisation parameter order and removed test settings for reset() function
Ultrasonic.py
Played with different sleep times between key reading and watchdog timer positions

Watchdog.py
Removed print statement from handler

Sonar.py
Initial commit. Spins a servo with and ultrasonic sensor on it and in future take the angle reading and distance measurment to form a 2D enviroment map of objects in front.
ISSUES; Ultrasonic sensor only works accuratly when an object is perpendicular to the sensor. Disstance is very inaccurate when object is at an angle.
Will return to this class in future and will try different distance sensors for better accuracy and reliaility.